Move i2 content (checker and tests) to `POSTabEligibilityChecker` and `POSTabEligibilityCheckerTests`.
