+++ WooCommerce/WooCommerceTests/ViewRelated/Settings/POS/POSTabEligibilityCheckerTests.swift
@@ -24,75 +24,15 @@ struct POSTabEligibilityCheckerTests {
         siteSettings = MockSelectedSiteSettings()
     }
 
-    @Test(arguments: [true, false])
-    func is_eligible_when_all_conditions_satisfied(isPointOfSaleAsATabi2Enabled: Bool) async throws {
-        // Given
-        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: isPointOfSaleAsATabi2Enabled)
-        setupCountry(country: .us)
-        accountWhitelistedInBackend(true)
-        let checker = POSTabEligibilityChecker(siteID: siteID,
-                                               userInterfaceIdiom: .pad,
-                                               siteSettings: siteSettings,
-                                               pluginsService: pluginsService,
-                                               stores: stores,
-                                               featureFlagService: featureFlagService)
-
-        // When
-        let result = await checker.checkEligibility()
-
-        // Then
-        #expect(result == .eligible)
-    }
-
-    @Test(arguments: [true, false])
-    func is_ineligible_when_account_not_whitelisted_and_feature_flag_disabled(isPointOfSaleAsATabi2Enabled: Bool) async throws {
-        // Given
-        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: isPointOfSaleAsATabi2Enabled)
-        setupCountry(country: .us)
-        accountWhitelistedInBackend(false)
-        let checker = POSTabEligibilityChecker(siteID: siteID,
-                                               userInterfaceIdiom: .pad,
-                                               siteSettings: siteSettings,
-                                               pluginsService: pluginsService,
-                                               stores: stores,
-                                               featureFlagService: featureFlagService)
-
-        // When
-        let result = await checker.checkEligibility()
-
-        // Then
-        #expect(result == .ineligible(reason: .featureFlagDisabled))
-    }
-
-    @Test(arguments: [true, false])
-    func is_ineligible_when_device_is_not_iPad(isPointOfSaleAsATabi2Enabled: Bool) async throws {
-        // Given
-        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: isPointOfSaleAsATabi2Enabled)
-        setupCountry(country: .us)
-        accountWhitelistedInBackend(true)
-        let checker = POSTabEligibilityChecker(siteID: siteID,
-                                               userInterfaceIdiom: .phone,
-                                               siteSettings: siteSettings,
-                                               pluginsService: pluginsService,
-                                               stores: stores,
-                                               featureFlagService: featureFlagService)
-
-        // When
-        let result = await checker.checkEligibility()
-
-        // Then
-        #expect(result == .ineligible(reason: .notTablet))
-    }
+    // MARK: `checkVisibility`
 
     @Test(arguments: [
-        (country: Country.us, currency: CurrencyCode.USD, isPointOfSaleAsATabi2Enabled: true),
-        (country: Country.us, currency: CurrencyCode.USD, isPointOfSaleAsATabi2Enabled: false),
-        (country: Country.gb, currency: CurrencyCode.GBP, isPointOfSaleAsATabi2Enabled: true),
-        (country: Country.gb, currency: CurrencyCode.GBP, isPointOfSaleAsATabi2Enabled: false)
+        (country: Country.us, currency: CurrencyCode.USD),
+        (country: Country.gb, currency: CurrencyCode.GBP)
     ])
-    fileprivate func is_eligible_when_country_and_currency_supported(country: Country, currency: CurrencyCode, isPointOfSaleAsATabi2Enabled: Bool) async throws {
+    fileprivate func is_visible_when_all_conditions_satisfied(country: Country, currency: CurrencyCode) async throws {
         // Given
-        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: isPointOfSaleAsATabi2Enabled)
+        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: true)
         setupCountry(country: country, currency: currency)
         accountWhitelistedInBackend(true)
         let checker = POSTabEligibilityChecker(siteID: siteID,
@@ -103,21 +43,19 @@ struct POSTabEligibilityCheckerTests {
                                                featureFlagService: featureFlagService)
 
         // When
-        let result = await checker.checkEligibility()
+        let result = await checker.checkVisibility()
 
         // Then
-        #expect(result == .eligible)
+        #expect(result == true)
     }
 
     @Test(arguments: [
-        (country: Country.ca, currency: CurrencyCode.CAD, isPointOfSaleAsATabi2Enabled: true),
-        (country: Country.ca, currency: CurrencyCode.CAD, isPointOfSaleAsATabi2Enabled: false),
-        (country: Country.es, currency: CurrencyCode.EUR, isPointOfSaleAsATabi2Enabled: true),
-        (country: Country.es, currency: CurrencyCode.EUR, isPointOfSaleAsATabi2Enabled: false)
+        (country: Country.ca, currency: CurrencyCode.CAD),
+        (country: Country.es, currency: CurrencyCode.EUR)
     ])
-    fileprivate func is_ineligible_when_country_is_not_supported(country: Country, currency: CurrencyCode, isPointOfSaleAsATabi2Enabled: Bool) async throws {
+    fileprivate func is_invisible_when_country_is_not_supported(country: Country, currency: CurrencyCode) async throws {
         // Given
-        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: isPointOfSaleAsATabi2Enabled)
+        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: true)
         setupCountry(country: country, currency: currency)
         accountWhitelistedInBackend(true)
         let checker = POSTabEligibilityChecker(siteID: siteID,
@@ -128,28 +66,22 @@ struct POSTabEligibilityCheckerTests {
                                                featureFlagService: featureFlagService)
 
         // When
-        let result = await checker.checkEligibility()
+        let result = await checker.checkVisibility()
 
         // Then
-        #expect(result == .ineligible(reason: .unsupportedCountry(supportedCountries: [.US, .GB])))
+        #expect(result == false)
     }
 
+
     @Test(arguments: [
-        (country: Country.us, currency: CurrencyCode.GBP, expectedSupportedCurrencies: [CurrencyCode.USD], isPointOfSaleAsATabi2Enabled: true),
-        (country: Country.us, currency: CurrencyCode.GBP, expectedSupportedCurrencies: [CurrencyCode.USD], isPointOfSaleAsATabi2Enabled: false),
-        (country: Country.us, currency: CurrencyCode.CAD, expectedSupportedCurrencies: [CurrencyCode.USD], isPointOfSaleAsATabi2Enabled: true),
-        (country: Country.us, currency: CurrencyCode.CAD, expectedSupportedCurrencies: [CurrencyCode.USD], isPointOfSaleAsATabi2Enabled: false),
-        (country: Country.gb, currency: CurrencyCode.EUR, expectedSupportedCurrencies: [CurrencyCode.GBP], isPointOfSaleAsATabi2Enabled: true),
-        (country: Country.gb, currency: CurrencyCode.EUR, expectedSupportedCurrencies: [CurrencyCode.GBP], isPointOfSaleAsATabi2Enabled: false),
-        (country: Country.gb, currency: CurrencyCode.USD, expectedSupportedCurrencies: [CurrencyCode.GBP], isPointOfSaleAsATabi2Enabled: true),
-        (country: Country.gb, currency: CurrencyCode.USD, expectedSupportedCurrencies: [CurrencyCode.GBP], isPointOfSaleAsATabi2Enabled: false)
+        (country: Country.us, currency: CurrencyCode.GBP),
+        (country: Country.us, currency: CurrencyCode.CAD),
+        (country: Country.gb, currency: CurrencyCode.EUR),
+        (country: Country.gb, currency: CurrencyCode.USD)
     ])
-    fileprivate func is_ineligible_when_currency_is_not_supported(country: Country,
-                                                                  currency: CurrencyCode,
-                                                                  expectedSupportedCurrencies: [CurrencyCode],
-                                                                  isPointOfSaleAsATabi2Enabled: Bool) async throws {
+    fileprivate func is_visible_when_currency_is_not_supported(country: Country, currency: CurrencyCode) async throws {
         // Given
-        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: isPointOfSaleAsATabi2Enabled)
+        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: true)
         setupCountry(country: country, currency: currency)
         accountWhitelistedInBackend(true)
         let checker = POSTabEligibilityChecker(siteID: siteID,
@@ -160,16 +92,16 @@ struct POSTabEligibilityCheckerTests {
                                                featureFlagService: featureFlagService)
 
         // When
-        let result = await checker.checkEligibility()
+        let result = await checker.checkVisibility()
 
         // Then
-        #expect(result == .ineligible(reason: .unsupportedCurrency(supportedCurrencies: expectedSupportedCurrencies)))
+        #expect(result == true)
     }
 
-    @Test(arguments: [true, false])
-    func is_ineligible_when_woocommerce_version_is_below_minimum(isPointOfSaleAsATabi2Enabled: Bool) async throws {
+
+    func is_visible_when_woocommerce_version_is_below_minimum() async throws {
         // Given
-        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: isPointOfSaleAsATabi2Enabled)
+        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: true)
         setupCountry(country: .us)
         accountWhitelistedInBackend(true)
         setupWooCommerceVersion("9.5.0")
@@ -181,16 +113,15 @@ struct POSTabEligibilityCheckerTests {
                                                featureFlagService: featureFlagService)
 
         // When
-        let result = await checker.checkEligibility()
+        let result = await checker.checkVisibility()
 
         // Then
-        #expect(result == .ineligible(reason: .unsupportedWooCommerceVersion(minimumVersion: "9.6.0-beta")))
+        #expect(result == true)
     }
 
-    @Test(arguments: [true, false])
-    func is_eligible_when_core_version_is_10_0_0_and_POS_feature_enabled(isPointOfSaleAsATabi2Enabled: Bool) async throws {
+    @Test func is_visible_when_core_version_is_10_0_0_and_POS_feature_enabled() async throws {
         // Given
-        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: isPointOfSaleAsATabi2Enabled)
+        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: true)
         setupCountry(country: .us)
         accountWhitelistedInBackend(true)
         setupWooCommerceVersion("10.0.0")
@@ -203,16 +134,16 @@ struct POSTabEligibilityCheckerTests {
                                                featureFlagService: featureFlagService)
 
         // When
-        let result = await checker.checkEligibility()
+        let result = await checker.checkVisibility()
 
         // Then
-        #expect(result == .eligible)
+        #expect(result == true)
     }
 
-    @Test(arguments: [true, false])
-    func is_ineligible_when_core_version_is_10_0_0_and_POS_feature_disabled(isPointOfSaleAsATabi2Enabled: Bool) async throws {
+
+    func is_visible_when_core_version_is_10_0_0_and_POS_feature_disabled() async throws {
         // Given
-        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: isPointOfSaleAsATabi2Enabled)
+        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: true)
         setupCountry(country: .us)
         accountWhitelistedInBackend(true)
         setupWooCommerceVersion("10.0.0")
@@ -225,16 +156,16 @@ struct POSTabEligibilityCheckerTests {
                                                featureFlagService: featureFlagService)
 
         // When
-        let result = await checker.checkEligibility()
+        let result = await checker.checkVisibility()
 
         // Then
-        #expect(result == .ineligible(reason: .featureSwitchDisabled))
+        #expect(result == true)
     }
 
-    @Test(arguments: [true, false])
-    func is_ineligible_when_core_version_is_10_0_0_and_POS_feature_check_fails(isPointOfSaleAsATabi2Enabled: Bool) async throws {
+
+    func is_visible_when_core_version_is_10_0_0_and_POS_feature_check_fails() async throws {
         // Given
-        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: isPointOfSaleAsATabi2Enabled)
+        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: true)
         setupCountry(country: .us)
         accountWhitelistedInBackend(true)
         setupWooCommerceVersion("10.0.0")
@@ -247,20 +178,58 @@ struct POSTabEligibilityCheckerTests {
                                                featureFlagService: featureFlagService)
 
         // When
-        let result = await checker.checkEligibility()
+        let result = await checker.checkVisibility()
 
         // Then
-        #expect(result == .ineligible(reason: .featureSwitchSyncFailure))
+        #expect(result == true)
     }
 
-    @Test(arguments: [true, false])
-    func is_eligible_when_core_version_is_below_10_0_0_and_POS_feature_disabled(isPointOfSaleAsATabi2Enabled: Bool) async throws {
+    @Test func is_visible_when_core_version_is_below_10_0_0_and_POS_feature_disabled() async throws {
         // Given
-        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: isPointOfSaleAsATabi2Enabled)
+        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: true)
         setupCountry(country: .us)
         accountWhitelistedInBackend(true)
         setupWooCommerceVersion("9.9.9")
         setupPOSFeatureEnabled(.success(false))
+        let checker = POSTabEligibilityChecker(siteID: siteID,
+                                               userInterfaceIdiom: .pad,
+                                               siteSettings: siteSettings,
+                                               pluginsService: pluginsService,
+                                               stores: stores,
+                                               featureFlagService: featureFlagService)
+
+        // When
+        let result = await checker.checkVisibility()
+
+        // Then
+        #expect(result == true)
+    }
+
+    @Test func is_visible_when_site_settings_are_from_correct_siteID() async throws {
+        // Given
+        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: true)
+
+        // Settings for a different site.
+        let wrongSiteSettings = [
+            mockCountrySetting(country: .ca, siteID: 999),
+            mockCurrencySetting(currency: .CAD, siteID: 999)
+        ]
+        // Settings for correct site.
+        let correctSiteSettings = [
+            mockCountrySetting(country: .us),
+            mockCurrencySetting(currency: .USD)
+        ]
+
+        siteSettings.mockSettingsStream = [
+            // Emits settings for a different site (should be filtered out).
+            (siteID: 999, settings: wrongSiteSettings, source: .storageChange),
+            // Emits first settings for correct site (should be skipped).
+            (siteID: siteID, settings: [SiteSetting.fake().copy(siteID: siteID, settingID: "temp")], source: .initialLoad),
+            // Emits fresh settings for correct site (should be used).
+            (siteID: siteID, settings: correctSiteSettings, source: .storageChange)
+        ].publisher.eraseToAnyPublisher()
+
+        accountWhitelistedInBackend(true)
         let checker = POSTabEligibilityChecker(siteID: siteID,
                                                userInterfaceIdiom: .pad,
                                                siteSettings: siteSettings,
@@ -274,6 +243,118 @@ struct POSTabEligibilityCheckerTests {
         // Then
         #expect(result == .eligible)
     }
+
+    @Test func is_invisible_when_remote_feature_flag_disabled() async throws {
+        // Given
+        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: true)
+        setupCountry(country: .us)
+        accountWhitelistedInBackend(false)
+        let checker = POSTabEligibilityChecker(siteID: siteID,
+                                               userInterfaceIdiom: .pad,
+                                               siteSettings: siteSettings,
+                                               pluginsService: pluginsService,
+                                               stores: stores,
+                                               featureFlagService: featureFlagService)
+
+        // When
+        let result = await checker.checkVisibility()
+
+        // Then
+        #expect(result == false)
+    }
+
+    @Test func checkVisibility_skips_settings_from_initialLoad() async throws {
+        // Given
+        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: true)
+
+        // Initial settings (cached) - makes site eligible (US)
+        let initialSettings = [
+            mockCountrySetting(country: .us),
+            mockCurrencySetting(currency: .USD)
+        ]
+        // New settings - makes site ineligible (Canada).
+        let newSettings = [
+            mockCountrySetting(country: .ca),
+            mockCurrencySetting(currency: .USD)
+        ]
+        siteSettings.mockSettingsStream = [
+            // Emits cached settings first (should be skipped).
+            (siteID: siteID, settings: initialSettings, source: .initialLoad),
+            // Emits new settings (should be used for eligibility check).
+            (siteID: siteID, settings: newSettings, source: .storageChange)
+        ].publisher.eraseToAnyPublisher()
+
+        accountWhitelistedInBackend(true)
+        let checker = POSTabEligibilityChecker(siteID: siteID,
+                                               userInterfaceIdiom: .pad,
+                                               siteSettings: siteSettings,
+                                               pluginsService: pluginsService,
+                                               stores: stores,
+                                               featureFlagService: featureFlagService)
+
+        // When
+        let result = await checker.checkVisibility()
+
+        // Then - Should be invisible because fresh settings show CA (not cached US)
+        #expect(result == false)
+    }
+
+    @Test func is_invisible_when_device_is_not_iPad() async throws {
+        // Given
+        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: true)
+        setupCountry(country: .us)
+        accountWhitelistedInBackend(true)
+        let checker = POSTabEligibilityChecker(siteID: siteID,
+                                               userInterfaceIdiom: .phone, // Not iPad
+                                               siteSettings: siteSettings,
+                                               pluginsService: pluginsService,
+                                               stores: stores,
+                                               featureFlagService: featureFlagService)
+
+        // When
+        let result = await checker.checkVisibility()
+
+        // Then
+        #expect(result == false)
+    }
+
+    @Test func checkVisibility_and_checkEligibility_return_expected_result_after_site_settings_available() async throws {
+        // Given - no site settings are immediately available (empty stream that will emit values later)
+        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: true)
+        accountWhitelistedInBackend(true)
+
+        // Creates a publisher that will emit values after a delay to simulate site settings loading
+        let countrySetting = mockCountrySetting(country: .us)
+        let currencySetting = mockCurrencySetting(currency: .USD)
+        let settingsSubject = PassthroughSubject<(siteID: Int64, settings: [SiteSetting], source: SettingsUpdateSource), Never>()
+        siteSettings.mockSettingsStream = settingsSubject.eraseToAnyPublisher()
+
+        let checker = POSTabEligibilityChecker(siteID: siteID,
+                                               userInterfaceIdiom: .pad,
+                                               siteSettings: siteSettings,
+                                               pluginsService: pluginsService,
+                                               stores: stores,
+                                               featureFlagService: featureFlagService)
+
+        // When - Call checkVisibility and checkEligibility concurrently before site settings are available
+        async let visibilityTask = checker.checkVisibility()
+        async let eligibilityTask = checker.checkEligibility()
+
+        // Simulate site settings becoming available after methods are called
+        Task {
+            settingsSubject.send((siteID: siteID, settings: [countrySetting, currencySetting], source: .refresh))
+            settingsSubject.send(completion: .finished)
+        }
+
+        let visibilityResult = await visibilityTask
+        let eligibilityResult = await eligibilityTask
+
+        // Then - both methods should wait for site settings and return expected results.
+        #expect(visibilityResult == true)
+        #expect(eligibilityResult == .eligible)
+    }
+
+    // MARK: - `checkInitialVisibility Tests
 
     @Test func checkInitialVisibility_returns_true_when_cached_tab_visibility_is_enabled() async throws {
         // Given
@@ -311,68 +392,16 @@ struct POSTabEligibilityCheckerTests {
         #expect(result == false)
     }
 
-    @Test(arguments: [true, false])
-    func checkEligibility_skips_settings_from_initialLoad(isPointOfSaleAsATabi2Enabled: Bool) async throws {
+    // MARK: - `checkEligibility` Tests
+
+    @Test(arguments: [
+        (country: Country.us, currency: CurrencyCode.USD),
+        (country: Country.gb, currency: CurrencyCode.GBP)
+    ])
+    fileprivate func is_eligible_when_all_conditions_satisfied(country: Country, currency: CurrencyCode) async throws {
         // Given
-        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: isPointOfSaleAsATabi2Enabled)
-
-        // Initial settings (cached) - makes site eligible (US)
-        let initialSettings = [
-            mockCountrySetting(country: .us),
-            mockCurrencySetting(currency: .USD)
-        ]
-        // New settings - makes site ineligible (Canada).
-        let newSettings = [
-            mockCountrySetting(country: .ca),
-            mockCurrencySetting(currency: .USD)
-        ]
-        siteSettings.mockSettingsStream = [
-            // Emits cached settings first (should be skipped).
-            (siteID: siteID, settings: initialSettings, source: .initialLoad),
-            // Emits new settings (should be used for eligibility check).
-            (siteID: siteID, settings: newSettings, source: .storageChange)
-        ].publisher.eraseToAnyPublisher()
-
-        accountWhitelistedInBackend(true)
-        let checker = POSTabEligibilityChecker(siteID: siteID,
-                                               userInterfaceIdiom: .pad,
-                                               siteSettings: siteSettings,
-                                               pluginsService: pluginsService,
-                                               stores: stores,
-                                               featureFlagService: featureFlagService)
-
-        // When
-        let result = await checker.checkEligibility()
-
-        // Then - Should be ineligible because fresh settings show CA (not cached US)
-        #expect(result == .ineligible(reason: .unsupportedCountry(supportedCountries: [.US, .GB])))
-    }
-
-    @Test(arguments: [true, false])
-    func checkEligibility_filters_by_correct_siteID(isPointOfSaleAsATabi2Enabled: Bool) async throws {
-        // Given
-        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: isPointOfSaleAsATabi2Enabled)
-
-        // Settings for a different site.
-        let wrongSiteSettings = [
-            mockCountrySetting(country: .ca, siteID: 999),
-            mockCurrencySetting(currency: .CAD, siteID: 999)
-        ]
-        // Settings for correct site.
-        let correctSiteSettings = [
-            mockCountrySetting(country: .us),
-            mockCurrencySetting(currency: .USD)
-        ]
-
-        siteSettings.mockSettingsStream = [
-            // Emits settings for a different site (should be filtered out).
-            (siteID: 999, settings: wrongSiteSettings, source: .storageChange),
-            // Emits first settings for correct site (should be skipped).
-            (siteID: siteID, settings: [SiteSetting.fake().copy(siteID: siteID, settingID: "temp")], source: .initialLoad),
-            // Emits fresh settings for correct site (should be used).
-            (siteID: siteID, settings: correctSiteSettings, source: .storageChange)
-        ].publisher.eraseToAnyPublisher()
-
+        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: true)
+        setupCountry(country: country, currency: currency)
         accountWhitelistedInBackend(true)
         let checker = POSTabEligibilityChecker(siteID: siteID,
                                                userInterfaceIdiom: .pad,
@@ -388,17 +417,38 @@ struct POSTabEligibilityCheckerTests {
         #expect(result == .eligible)
     }
 
-    // MARK: - checkVisibility Tests
+    @Test(arguments: [
+        (country: Country.ca, currency: CurrencyCode.CAD),
+        (country: Country.es, currency: CurrencyCode.EUR)
+    ])
+    fileprivate func is_ineligible_when_country_is_not_supported(country: Country, currency: CurrencyCode) async throws {
+        // Given
+        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: true)
+        setupCountry(country: country, currency: currency)
+        accountWhitelistedInBackend(true)
+        let checker = POSTabEligibilityChecker(siteID: siteID,
+                                               userInterfaceIdiom: .pad,
+                                               siteSettings: siteSettings,
+                                               pluginsService: pluginsService,
+                                               stores: stores,
+                                               featureFlagService: featureFlagService)
+
+        // When
+        let result = await checker.checkEligibility()
+
+        // Then
+        #expect(result == .ineligible(reason: .siteSettingsNotAvailable))
+    }
 
     @Test(arguments: [
-        // Eligible countries and currencies.
-        (country: Country.us, currency: CurrencyCode.USD),
-        (country: Country.gb, currency: CurrencyCode.GBP),
-        // Eligible countries but ineligible currencies.
-        (country: Country.us, currency: CurrencyCode.EUR),
-        (country: Country.gb, currency: CurrencyCode.CAD)
+        (country: Country.us, currency: CurrencyCode.GBP, expectedSupportedCurrencies: [CurrencyCode.USD]),
+        (country: Country.us, currency: CurrencyCode.CAD, expectedSupportedCurrencies: [CurrencyCode.USD]),
+        (country: Country.gb, currency: CurrencyCode.EUR, expectedSupportedCurrencies: [CurrencyCode.GBP]),
+        (country: Country.gb, currency: CurrencyCode.USD, expectedSupportedCurrencies: [CurrencyCode.GBP])
     ])
-    fileprivate func checkVisibility_returns_true_when_i2_enabled_and_country_remote_feature_eligible(country: Country, currency: CurrencyCode) async throws {
+    fileprivate func is_ineligible_when_currency_is_not_supported(country: Country,
+                                                                  currency: CurrencyCode,
+                                                                  expectedSupportedCurrencies: [CurrencyCode]) async throws {
         // Given
         let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: true)
         setupCountry(country: country, currency: currency)
@@ -411,116 +461,18 @@ struct POSTabEligibilityCheckerTests {
                                                featureFlagService: featureFlagService)
 
         // When
-        let result = await checker.checkVisibility()
+        let result = await checker.checkEligibility()
 
         // Then
-        #expect(result == true)
+        #expect(result == .ineligible(reason: .unsupportedCurrency(supportedCurrencies: expectedSupportedCurrencies)))
     }
 
-    @Test(arguments: [(country: Country.ca, currency: CurrencyCode.CAD), (country: Country.es, currency: CurrencyCode.EUR)])
-    fileprivate func checkVisibility_returns_false_when_pointOfSaleAsATabi2_enabled_but_country_ineligible(country: Country, currency: CurrencyCode) async throws {
-        // Given
-        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: true)
-        setupCountry(country: country, currency: currency)
-        accountWhitelistedInBackend(true)
-        let checker = POSTabEligibilityChecker(siteID: siteID,
-                                               userInterfaceIdiom: .pad,
-                                               siteSettings: siteSettings,
-                                               pluginsService: pluginsService,
-                                               stores: stores,
-                                               featureFlagService: featureFlagService)
-
-        // When
-        let result = await checker.checkVisibility()
-
-        // Then
-        #expect(result == false)
-    }
-
-    @Test(arguments: [(country: Country.us, currency: CurrencyCode.USD), (country: Country.gb, currency: .GBP)])
-    fileprivate func checkVisibility_returns_false_when_i2_enabled_but_remote_feature_flag_disabled(country: Country, currency: CurrencyCode) async throws {
-        // Given
-        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: true)
-        setupCountry(country: country, currency: currency)
-        accountWhitelistedInBackend(false)
-        let checker = POSTabEligibilityChecker(siteID: siteID,
-                                               userInterfaceIdiom: .pad,
-                                               siteSettings: siteSettings,
-                                               pluginsService: pluginsService,
-                                               stores: stores,
-                                               featureFlagService: featureFlagService)
-
-        // When
-        let result = await checker.checkVisibility()
-
-        // Then
-        #expect(result == false)
-    }
-
-    @Test func checkVisibility_returns_true_when_pointOfSaleAsATabi2_disabled_and_checkEligibility_eligible() async throws {
-        // Given
-        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: false)
-        setupCountry(country: .us)
-        accountWhitelistedInBackend(true)
-        let checker = POSTabEligibilityChecker(siteID: siteID,
-                                               userInterfaceIdiom: .pad,
-                                               siteSettings: siteSettings,
-                                               pluginsService: pluginsService,
-                                               stores: stores,
-                                               featureFlagService: featureFlagService)
-
-        // When
-        let result = await checker.checkVisibility()
-
-        // Then
-        #expect(result == true)
-    }
-
-    @Test(arguments: [(country: Country.us, currency: CurrencyCode.GBP), (country: Country.gb, currency: .EUR)])
-    fileprivate func checkVisibility_returns_false_when_i2_disabled_and_checkEligibility_ineligible(country: Country, currency: CurrencyCode) async throws {
-        // Given
-        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: false)
-        setupCountry(country: country, currency: currency) // Ineligible country/currency combination
-        accountWhitelistedInBackend(true)
-        let checker = POSTabEligibilityChecker(siteID: siteID,
-                                               userInterfaceIdiom: .pad,
-                                               siteSettings: siteSettings,
-                                               pluginsService: pluginsService,
-                                               stores: stores,
-                                               featureFlagService: featureFlagService)
-
-        // When
-        let result = await checker.checkVisibility()
-
-        // Then
-        #expect(result == false)
-    }
-
-    @Test(arguments: [true, false])
-    func checkVisibility_returns_false_when_device_is_not_iPad(isPointOfSaleAsATabi2Enabled: Bool) async throws {
-        // Given
-        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: isPointOfSaleAsATabi2Enabled)
-        setupCountry(country: .us)
-        accountWhitelistedInBackend(true)
-        let checker = POSTabEligibilityChecker(siteID: siteID,
-                                               userInterfaceIdiom: .phone, // Not iPad
-                                               siteSettings: siteSettings,
-                                               pluginsService: pluginsService,
-                                               stores: stores,
-                                               featureFlagService: featureFlagService)
-
-        // When
-        let result = await checker.checkVisibility()
-
-        // Then
-        #expect(result == false)
-    }
-
-    @Test func checkEligibility_uses_cached_values_after_checkVisibility_when_i2_feature_is_enabled() async throws {
+    func is_ineligible_when_woocommerce_version_is_below_minimum() async throws {
         // Given
         let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: true)
         setupCountry(country: .us)
         accountWhitelistedInBackend(true)
+        setupWooCommerceVersion("9.5.0")
         let checker = POSTabEligibilityChecker(siteID: siteID,
                                                userInterfaceIdiom: .pad,
                                                siteSettings: siteSettings,
@@ -528,32 +480,20 @@ struct POSTabEligibilityCheckerTests {
                                                stores: stores,
                                                featureFlagService: featureFlagService)
 
-        // When checkVisibility first (which caches siteSettingsEligibility and featureFlagEligibility)
-        let visibilityResult = await checker.checkVisibility()
+        // When
+        let result = await checker.checkEligibility()
 
-        // And site settings and feature flag eligibility changes
-        setupCountry(country: .ca, currency: .AMD)
-        accountWhitelistedInBackend(false)
-
-        // Then checkEligibility should use cached values for site settings and feature flags
-        let eligibilityResult = await checker.checkEligibility()
-
-        // Then - both should return the expected results, demonstrating caching works
-        #expect(visibilityResult == true)
-        #expect(eligibilityResult == .eligible)
+        // Then
+        #expect(result == .ineligible(reason: .unsupportedWooCommerceVersion(minimumVersion: "9.6.0-beta")))
     }
 
-    @Test func checkVisibility_and_checkEligibility_return_expected_result_after_site_settings_available() async throws {
-        // Given - no site settings are immediately available (empty stream that will emit values later)
+    func is_eligible_when_core_version_is_10_0_0_and_POS_feature_enabled() async throws {
+        // Given
         let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: true)
+        setupCountry(country: .us)
         accountWhitelistedInBackend(true)
-
-        // Creates a publisher that will emit values after a delay to simulate site settings loading
-        let countrySetting = mockCountrySetting(country: .us)
-        let currencySetting = mockCurrencySetting(currency: .USD)
-        let settingsSubject = PassthroughSubject<(siteID: Int64, settings: [SiteSetting], source: SettingsUpdateSource), Never>()
-        siteSettings.mockSettingsStream = settingsSubject.eraseToAnyPublisher()
-
+        setupWooCommerceVersion("10.0.0")
+        setupPOSFeatureEnabled(.success(true))
         let checker = POSTabEligibilityChecker(siteID: siteID,
                                                userInterfaceIdiom: .pad,
                                                siteSettings: siteSettings,
@@ -561,22 +501,74 @@ struct POSTabEligibilityCheckerTests {
                                                stores: stores,
                                                featureFlagService: featureFlagService)
 
-        // When - Call checkVisibility and checkEligibility concurrently before site settings are available
-        async let visibilityTask = checker.checkVisibility()
-        async let eligibilityTask = checker.checkEligibility()
+        // When
+        let result = await checker.checkEligibility()
 
-        // Simulate site settings becoming available after methods are called
-        Task {
-            settingsSubject.send((siteID: siteID, settings: [countrySetting, currencySetting], source: .refresh))
-            settingsSubject.send(completion: .finished)
-        }
+        // Then
+        #expect(result == .eligible)
+    }
 
-        let visibilityResult = await visibilityTask
-        let eligibilityResult = await eligibilityTask
+    func is_ineligible_when_core_version_is_10_0_0_and_POS_feature_disabled() async throws {
+        // Given
+        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: true)
+        setupCountry(country: .us)
+        accountWhitelistedInBackend(true)
+        setupWooCommerceVersion("10.0.0")
+        setupPOSFeatureEnabled(.success(false))
+        let checker = POSTabEligibilityChecker(siteID: siteID,
+                                               userInterfaceIdiom: .pad,
+                                               siteSettings: siteSettings,
+                                               pluginsService: pluginsService,
+                                               stores: stores,
+                                               featureFlagService: featureFlagService)
 
-        // Then - both methods should wait for site settings and return expected results.
-        #expect(visibilityResult == true)
-        #expect(eligibilityResult == .eligible)
+        // When
+        let result = await checker.checkEligibility()
+
+        // Then
+        #expect(result == .ineligible(reason: .featureSwitchDisabled))
+    }
+
+    func is_ineligible_when_core_version_is_10_0_0_and_POS_feature_check_fails() async throws {
+        // Given
+        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: true)
+        setupCountry(country: .us)
+        accountWhitelistedInBackend(true)
+        setupWooCommerceVersion("10.0.0")
+        setupPOSFeatureEnabled(.failure(NSError(domain: "test", code: 0)))
+        let checker = POSTabEligibilityChecker(siteID: siteID,
+                                               userInterfaceIdiom: .pad,
+                                               siteSettings: siteSettings,
+                                               pluginsService: pluginsService,
+                                               stores: stores,
+                                               featureFlagService: featureFlagService)
+
+        // When
+        let result = await checker.checkEligibility()
+
+        // Then
+        #expect(result == .ineligible(reason: .featureSwitchSyncFailure))
+    }
+
+    func is_eligible_when_core_version_is_below_10_0_0_and_POS_feature_disabled() async throws {
+        // Given
+        let featureFlagService = MockFeatureFlagService(isPointOfSaleAsATabi2Enabled: true)
+        setupCountry(country: .us)
+        accountWhitelistedInBackend(true)
+        setupWooCommerceVersion("9.9.9")
+        setupPOSFeatureEnabled(.success(false))
+        let checker = POSTabEligibilityChecker(siteID: siteID,
+                                               userInterfaceIdiom: .pad,
+                                               siteSettings: siteSettings,
+                                               pluginsService: pluginsService,
+                                               stores: stores,
+                                               featureFlagService: featureFlagService)
+
+        // When
+        let result = await checker.checkEligibility()
+
+        // Then
+        #expect(result == .eligible)
     }
 }
